fix(p3-m1): audit hardening — wedge epsilon, shadow-fit stability, batching
Closing double audit (csharp-lowlevel + engine-architect): PASS/PASS, no
blocker, two majors fixed in-milestone.

- ExtrudedShadowFrustum: the epsilon rule dropped the planes EXACTLY parallel
  to the light ray — the very planes that close the wedge sideways. With the
  sun overhead and a level camera the four lateral planes fell and the wedge
  culled nothing; the bench escaped it only because its light is axis-unaligned.
  Drop a plane iff dot(n, dir) > +eps. Adds a zenith test; rewrites the
  near-parallel test, which pinned the wrong behaviour.
- ShadowFit: the scene branch skipped the texel snap on the grounds that "a
  static scene cannot shimmer" — an assumption killed by per-frame world bounds.
  Quantize the radius (16 rungs per octave) and snap the centre in BOTH
  branches, so the fit is a step function of the scene and shadow edges stop
  crawling. The helmet capture is therefore no longer byte-identical to phase 2
  (0.25% of channels, sub-texel shadow shift); render verified intact.
- Shadow list carries a MESH-major key: the depth pass binds no material, so a
  mesh shared by k materials no longer fragments into k draws.
- InstanceBufferRing: lifts the duplicated compaction out of Renderer and shrinks
  a slot's SSBO after 60 frames below a quarter of its capacity.
- Rebind set 1 only when the material changes; declare StorageBuffer in the
  persistent descriptor pool; 64-bit multiply in GpuBuffer.Write; document the
  16-bit mesh/material ceiling as a hard limit.

Gates: 284 tests, 0 warning, 0 validation message, 0 leak, 0 alloc/frame,
NativeAOT PASS (10k bench: 2 draw calls, ~2.2 ms cull+collect).

+++ src/Agapanthe.Graphics/GpuBuffer.cs
@@ -120,7 +120,9 @@ public sealed unsafe class GpuBuffer : IDisposable
                 "Write<T> is only valid on a host-visible buffer; use the staging upload path for device-local memory.");
         }
 
-        var bytes = (ulong)(data.Length * Unsafe.SizeOf<T>());
+        // 64-bit multiply: a span of > 2 GiB worth of elements would overflow an int product and report a nonsense
+        // size (the check below would still reject it, with an absurd message).
+        var bytes = (ulong)((long)data.Length * Unsafe.SizeOf<T>());
         if (bytes > SizeBytes)
         {
             throw new GraphicsException($"Write of {bytes} bytes exceeds buffer size {SizeBytes}.");
